a bunch more script changes and folder structure also draw rope coroutine
