+++ Assets/FunctionalTeams/Player/Scripts/CinemachineShake.cs
