+++ Assets/FunctionalTeams/Player/Scripts/CinemachineShake.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 343dc56b5628ad7418c196d4006c3170
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
