+++ Assets/FunctionalTeams/Player/Scripts/Harpoon.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06961d21394dbf344935990017a857f5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
